Share provider-neutral OpenTelemetry configuration
