+++ src/Temporalio.Extensions.Aws.Lambda.OpenTelemetry/Temporalio.Extensions.Aws.Lambda.OpenTelemetry.csproj
@@ -11,8 +11,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="OpenTelemetry" />
-    <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" />
     <PackageReference Include="OpenTelemetry.Extensions.AWS" />
   </ItemGroup>
