+++ tests/Temporalio.Tests/Extensions/OpenTelemetry/NoopClientInterceptor.cs
@@ -1,0 +1,9 @@
+namespace Temporalio.Tests.Extensions.OpenTelemetry;
+
+using Temporalio.Client.Interceptors;
+
+internal sealed class NoopClientInterceptor : IClientInterceptor
+{
+    public ClientOutboundInterceptor InterceptClient(
+        ClientOutboundInterceptor nextInterceptor) => nextInterceptor;
+}
