+++ tests/Temporalio.Tests/Extensions/OpenTelemetry/OpenTelemetryEnvironmentDefinition.cs
@@ -1,0 +1,9 @@
+namespace Temporalio.Tests.Extensions.OpenTelemetry;
+
+using Xunit;
+
+[CollectionDefinition(Name, DisableParallelization = true)]
+public sealed class OpenTelemetryEnvironmentDefinition
+{
+    public const string Name = "OpenTelemetryEnvironment";
+}
